cic(#1107): lift the quotable-body rules out of the reply wrapper
`replyQuote` had grown into two things: the rules for deciding whether a
scrollback row has anything quotable in it and, if so, extracting the
words the operator saw — and the `<nick> …<< ` wrapper reply puts around
them. #1107 needs the first half with a different wrapper.

Every clause in that first half is a ruling that cost an issue to reach:
#1067's content-kind gate (a PART's reason lives in `body`, so a bare
body check quotes "Leaving"), #1123's previous-quote cut and the nick
charset it is anchored on, #1126's CTCP unwrap. Copying them into a
second module would copy five bugs' worth of history and let the two
doors drift on the sixth, so they move to `lib/quotableBody.ts` and both
callers share them.

`quotableBody` rather than an export from `replyQuote`: a module named
for one caller is the wrong home for a rule both obey, and importing
"reply" from an addquote path is a name that lies.

No behaviour change — `replyQuote`'s 26 existing assertions pass
unmodified, which is what makes this a lift and not a rewrite.
